# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book-Store Backend API",
    "description": "Documentation for our Express/MongoDB backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8888/api/"
    }
  ],
  "paths": {
...
    "/cart/": {
      "get": {
        "tags": [
          "Cart"
        ],
        "summary": "Get current user cart",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Add/Update item in cart",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cartItemSchema"
              }
            }
          }
        }
      }
    },
    "/cart/{bookId}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "Remove item from cart",
        "description": "",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Register a new user",
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userSchema"
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login and receive JWT",
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginSchema"
              }
            }
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get current logged-in user profile",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update current user profile",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateUserSchema"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "cartItemSchema": {
        "type": "object",
        "properties": {
          "book": {
            "type": "string",
            "pattern": "^[0-9a-f]{24}$/"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "book",
          "quantity"
        ],
        "additionalProperties": false
      },
...
      "userSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "title": "First Name"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "title": "Last Name"
          },
          "dob": {
            "type": "string",
            "format": "date-time",
            "title": "Date of birth"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "dob",
          "password"
        ],
        "additionalProperties": false
      },
      "loginSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "firstName": {},
          "lastName": {},
          "dob": {},
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "updateUserSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "title": "First Name"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "title": "Last Name"
          },
          "dob": {
            "type": "string",
            "format": "date-time",
            "title": "Date of birth"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}